third_party/WebKit/LayoutTests/inspector/elements/styles-3/selector-source-data.html

<html>
<head>

<style>
/* c1 */
                         html
  #inspected.bar /* c2 */,
 /* c3 */ b
  /* c4 */  {
    text-decoration: none;
}
</style>

<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/elements-test.js"></script>
<script>

function test()
{
    InspectorTest.nodeWithId("inspected", nodeFound);

    async function nodeFound(node)
    {
        var response = await InspectorTest.CSSAgent.invoke_getMatchedStylesForNode({nodeId: node.id});
        if (response[Protocol.Error]) {
            InspectorTest.addResult("Failed to get styles: " + response[Protocol.Error]);
            return;
        }
        InspectorTest.dumpRuleMatchesArray(response.matchedCSSRules);
        InspectorTest.completeTest();
    }
}
</script>
</head>

<body onload="runTest()">
<p>
Tests that WebInspector.CSSStyleSheet methods work as expected.
</p>
<h1 id="inspected" class="bar">Inspect Me</h1>
</body>
</html>
