third_party/WebKit/LayoutTests/inspector/elements/styles-3/selector-source-data-expected.txt

Tests that WebInspector.CSSStyleSheet methods work as expected.

Inspect Me

h1: [user-agent] {
    ['display':'block'] @[undefined-undefined] 
    ['font-size':'2em'] @[undefined-undefined] 
    ['-webkit-margin-before':'0.67em'] @[undefined-undefined] 
    ['-webkit-margin-after':'0.67em'] @[undefined-undefined] 
    ['-webkit-margin-start':'0px'] @[undefined-undefined] 
    ['-webkit-margin-end':'0px'] @[undefined-undefined] 
    ['font-weight':'bold'] @[undefined-undefined] 
}
html #inspected.bar, b: [regular, 2:25-3:25; 4:10-5:10] {
    ['text-decoration':'none'] @[6:4-6:26] 
    ['text-decoration-line':'none'] @[undefined-undefined] 
    ['text-decoration-style':'initial'] @[undefined-undefined] 
    ['text-decoration-color':'initial'] @[undefined-undefined] 
}
